// src/types/user.ts
export interface UserProfile {
  id?: string;
  username?: string;
  description?: string;
  profileImage?: string;
  profile_image?: string;
  social_links?: {
    instagram?: string;
    youtube?: string;
    website?: string;
  };
  socialLinks?: {
    instagram?: string;
    youtube?: string;
    website?: string;
  };
}